requirements.txt
albumentations>=1.4.0
matplotlib==3.10.8
numpy==2.4.3
Pillow==12.1.1
torch==2.8.0
torchinfo==1.8.0
torchvision==0.23.0
tqdm==4.67.1
